+++ src/main/java/org/kwakmunsu/flowmate/domain/challenge/entity/dto/ChallengeCreateDomainRequest.java
@@ -12,7 +12,7 @@ public record ChallengeCreateDomainRequest(
         LocalDate startDate,
         LocalDate endDate,
         String rule,
-        Integer maxParticipants
+        Long maxParticipants
 ) {
 
 }
